Aruco Tags and Pose Estimation
Mission of this project is to use computer vision to recognize and identify the tags, and display the location of the “on/off” switch to the user. The markers are 4x4 bit patterns, size is 2 inches on a side, and the markers were created by specifying 100 markers in the dictionary. The marker on the left is id=0; the marker on the right is id=1.

Here is the location of the switch (in inches):
With respect to marker 0: (X,Y,Z) = (2.5; -2.0; -1.0).
With respect to marker 1: (X,Y,Z) = (-2.5; -2.0; -5.0).

This program will recognize a marker, determine its pose, and overlay a circle on the image showing the location of the on/off switch as an overlay.