Orthophoto / Detection
The video “checkers.mp4”, It is a short clip that shows a game of checkers. Process every 10th image from the video. In each of these images, try to recognize the board and find the pieces. Note that you may not be able to find the board in each image. Even if you find the board, you may not correctly detect all the pieces. However, by processing all the images in the clip, we can accumulate all the detections and get a better idea of where they are.

After processing the entire clip, record the number of times that you detected a white piece and the number of times you detected a black piece in each of the 8x8 squares. Pick a threshold such that if the count is equal to or higher than the threshold, then you decide that the square contains the appropriate color piece.